fix: move tracks components to questing domain
