+++ apps/govquests-frontend/src/domains/questing/components/track/TrackAccordionHeader.tsx
@@ -1,4 +1,4 @@
-import { Tracks } from "../types/trackTypes";
+import { Tracks } from "../../types/trackTypes";
 
 interface TrackAccordionHeader {
   track: Tracks[number];
