+++ apps/govquests-frontend/src/domains/questing/components/track/TrackList.tsx
@@ -1,4 +1,4 @@
-import { useFetchTracks } from "../hooks/useFetchTracks";
+import { useFetchTracks } from "../../hooks/useFetchTracks";
 import { TrackAccordion } from "./TrackAccordion";
 
 export const TrackList = () => {
